src/chatty/pyproject.toml
[tool.poetry]
name = "remofirst-ag-techctask"
version = "1.0.0"
description = "Real-time WebSocket protocol layer based Django REST backend application."
authors = ["arifgarayev <garayevarif@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytype = "^2024.4.11"
flit = "3.8.0"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
gunicorn = "^22.0.0"
daphne = "^4.1.2"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
django-debug-toolbar = "^4.3.0"
pre-commit = "^3.7.0"
celery = "^5.4.0"
pydantic = "^2.7.1"
coreapi = "^2.3.3"
bcrypt = "^4.1.2"
django-rest-swagger = "^2.2.0"
drf-spectacular = "^0.27.2"
channels = "^4.1.0"
kafka-python = "^2.0.2"
gevent-websocket = "^0.10.1"
sqlalchemy = "^2.0.29"
websocket-client = "^1.8.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
flake8 = "^7.0.0"
Flake8-pyproject = "^1.2.3"
bandit = "^1.7.8"
flake8-bugbear = "^24.4.21"
flake8-cognitive-complexity = "^0.1.0"
pre-commit = "^3.7.0"
safety = "^3.1.0"
pip-audit = "^2.7.2"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 140
py_version = 311
order_by_type = false
skip = [".gitignore", ".dockerignore"]
extend_skip = [".md", ".json"]
skip_glob = ["docs/*"]

[tool.flake8]
max-line-length = 140
select = ["C", "E", "F", "W", "B", "B9"]
ignore = ["E203", "E501", "W503", "C812", "E731", "F811"]
exclude = ["__init__.py"]

[tool.black]
line-length = 140
target-version = ['py311']
include = '\.pyi?$'

[tool.bandit]
skips = ["B311", "B404"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration relatively slow (deselect with '-m \"not integration\"')",
    "serial",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]

testpaths = "tests"